+++ src/CommandLine/Program.cs
@@ -640,17 +640,32 @@ internal static class Program
         if (!options.TryGetProjectFilter(out ProjectFilter projectFilter))
             return ExitCodes.Error;
 
-        if (!ParseHelpers.TryEnsureFullPath(options.Words, out ImmutableArray<string> wordListPaths))
+        if (!TryEnsureFullPath(options.Words, out ImmutableArray<string> wordListPaths))
             return ExitCodes.Error;
+
+        foreach (string path in wordListPaths)
+        {
+            if (!File.Exists(path)
+                && !Directory.Exists(path))
+            {
+                WriteLine($"File or directory not found: '{path}'.", ConsoleColors.Yellow, Verbosity.Quiet);
+                return ExitCodes.Error;
+            }
+        }
 
         if (!TryParsePaths(options.Paths, out ImmutableArray<string> paths))
             return ExitCodes.Error;
+
+        var loadOptions = WordListLoadOptions.DetectNonWords;
+
+        if (!options.CaseSensitive)
+            loadOptions |= WordListLoadOptions.IgnoreCase;
 
         WordListLoaderResult loaderResult = WordListLoader.Load(
             wordListPaths,
             options.MinWordLength,
             options.MaxWordLength,
-            (options.CaseSensitive) ? WordListLoadOptions.None : WordListLoadOptions.IgnoreCase);
+            loadOptions);
 
         var data = new SpellingData(loaderResult.List, loaderResult.CaseSensitiveList, loaderResult.FixList);