+++ src/Workspaces.Core/Spelling/Core/WordCharMap.cs
@@ -45,7 +45,7 @@ internal sealed class WordCharMap
 
     public static WordCharMap CreateCharIndexMap(WordList wordList, bool reverse = false)
     {
-        ImmutableDictionary<WordChar, ImmutableHashSet<string>> map = wordList.Values
+        ImmutableDictionary<WordChar, ImmutableHashSet<string>> map = wordList.Words
             .Select(s =>
             {
                 return (